Implements initial pass of learn2code heuristic filters (#230)
This is a combination of @revbucket 's desired heuristic filters for learn2code,
plus the existing filters we had previously implemented in prior iterations.

A starting point. I am a bit suspicious about the precision of the regexes
I lifted from the StarCoder paper.

Note: plan is to merge this into long-lived "learn2code"
feature branch rather than main, as I suspect we will be iterating
a lot.