Code-prose-composition tagger.
Add a tagger that adds attributes for code-prose-other
composition of files based on line classifications.

Coverage for code-prose-composition tagger.

Improve error messages for spawn method checks.

Set multiprocessing start method in test setup

Set multiprocessing in test case with error handling.

Add before suite hook to set mp start method.

The default multiprocessing start method is "fork" which is not compatible with
with runtime assertions that it is set to spawn. When running unit tests, it's
possible to call an external library that sets the start method to "fork".
Here we enforce the start method to be "spawn" for all tests before executing.

linting.

Remove error log messages.